% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratToMonocle.R
\name{SeuToMon}
\alias{SeuToMon}
\title{SeuToMon}
\usage{
SeuToMon(seuObj, str_assay)
}
\arguments{
\item{seuObj}{The SeuratObject to convert}
\item{str_assay}{The assay of the SeuratObject to use as expression matrix}
}
\value{
Cell DataSet Objects
}
\description{
SeuToMon
}

